# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OmniCreator AI Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List projects in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: archived
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a new project
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectBody'
      responses:
        '201':
          description: Created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{projectId}:
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get project details
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Update project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectBody'
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Delete project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /projects/{projectId}/duplicate:
    post:
      tags: [Projects]
      operationId: duplicateProject
      summary: Duplicate a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Duplicated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{projectId}/archive:
    post:
      tags: [Projects]
      operationId: archiveProject
      summary: Archive a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{projectId}/restore:
    post:
      tags: [Projects]
      operationId: restoreProject
      summary: Restore an archived project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

  # VIDEO EDITOR
  /projects/{projectId}/timeline:
    get:
      tags: [VideoEditor]
      operationId: getTimeline
      summary: Get video editor timeline state
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Timeline document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineDocument'
    put:
      tags: [VideoEditor]
      operationId: saveTimeline
      summary: Save video editor timeline state
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveTimelineBody'
      responses:
        '200':
          description: Saved timeline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineDocument'

  # IMAGE EDITOR
  /projects/{projectId}/canvas:
    get:
      tags: [ImageEditor]
      operationId: getCanvas
      summary: Get image editor canvas state
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Canvas document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasDocument'
    put:
      tags: [ImageEditor]
      operationId: saveCanvas
      summary: Save image editor canvas state
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveCanvasBody'
      responses:
        '200':
          description: Saved canvas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasDocument'

  # ASSETS
...
components:
...
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # AUTH
...
    Project:
      type: object
      required: [id, workspaceId, name, type, status, createdById, createdAt, updatedAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [video, image, carousel, ad, thumbnail, product_mockup, campaign, template, asset_only]
        status:
          type: string
          enum: [draft, in_progress, review, approved, published, archived]
        platform:
          type: ["string", "null"]
        aspectRatio:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
        language:
          type: ["string", "null"]
        brandKitId:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        isArchived:
          type: boolean
        createdById:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    CreateProjectBody:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        platform:
          type: string
        aspectRatio:
          type: string
        duration:
          type: number
        language:
          type: string
        brandKitId:
          type: string
        tags:
          type: array
          items:
            type: string
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        platform:
          type: ["string", "null"]
        aspectRatio:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
        language:
          type: ["string", "null"]
        brandKitId:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string

    # VIDEO EDITOR
    TimelineDocument:
      type: object
      required: [id, projectId, version, durationMs, updatedAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        version:
          type: integer
        durationMs:
          type: number
        tracks:
          type: array
          items:
            type: object
        captions:
          type: array
          items:
            type: object
        audioMix:
          type: object
        exportSettings:
          type: object
        updatedAt:
          type: string
    SaveTimelineBody:
      type: object
      required: [tracks, durationMs]
      properties:
        durationMs:
          type: number
        tracks:
          type: array
          items:
            type: object
        captions:
          type: array
          items:
            type: object
        audioMix:
          type: object
        exportSettings:
          type: object

    # IMAGE EDITOR
    CanvasDocument:
      type: object
      required: [id, projectId, version, updatedAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        version:
          type: integer
        pages:
          type: array
          items:
            type: object
        sharedAssets:
          type: array
          items:
            type: string
        brandKitId:
          type: ["string", "null"]
        updatedAt:
          type: string
    SaveCanvasBody:
      type: object
      required: [pages]
      properties:
        pages:
          type: array
          items:
            type: object
        sharedAssets:
          type: array
          items:
            type: string
        brandKitId:
          type: ["string", "null"]

    # ASSETS
...