# packages/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OmniCreator AI Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectBody {
  name: string;
  type: string;
  platform?: string;
  aspectRatio?: string;
  duration?: number;
  language?: string;
  brandKitId?: string;
  tags?: string[];
}