# packages/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OmniCreator AI Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProjectBody {
  name?: string;
  status?: string;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  aspectRatio?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  brandKitId?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  tags?: string[];
}